requirements_cpu.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
compressed-rtf==1.0.6
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.29.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.3.0
gliner==0.2.16
gradio==5.20.1
gradio-client==1.7.2
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imapclient==2.1.0
jinja2==3.1.6
jiter==0.9.0
lxml==5.3.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
olefile==0.47
onnxruntime==1.21.0
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfminer-six==20191110
pillow==11.1.0
pip==24.0
protobuf==6.30.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.10.0
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==76.0.0
shellingham==1.5.4
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
speechrecognition==3.8.1
starlette==0.46.1
sympy==1.13.1
textract==1.6.5
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
word2number==1.1
xlrd==1.2.0
xlsxwriter==3.2.2
